Rework and test network.rs

+++ src/block/page.rs
@@ -40,7 +40,7 @@ pub struct PageHeader
     pub site_id: Signature,
 
     pub page_name: String,
-    pub page_fee: f64,
+    pub page_fee: f32,
 }
 
 impl PageHeader
@@ -86,7 +86,7 @@ impl ToString for Page
 impl Page
 {
 
-    pub fn new(format: DataFormat, data: Vec<u8>, owner: &PrivateWallet, page_name: &str, fee: f64) -> Self
+    pub fn new(format: DataFormat, data: Vec<u8>, owner: &PrivateWallet, page_name: &str, fee: f32) -> Self
     {
         let header = PageHeader
         {
@@ -107,7 +107,7 @@ impl Page
     }
 
     /*
-    pub fn from_file(chain: &BlockChain, new_page: &[u8], owner: &PrivateWallet, page_name: &str, fee: f64) -> Self
+    pub fn from_file(chain: &BlockChain, new_page: &[u8], owner: &PrivateWallet, page_name: &str, fee: f32) -> Self
     {
         let existing_page_file = chain.page(owner, page_name);
         if existing_page_file.is_none() {
